+++ server/src/admin/api-docs/loginAdmin.api-docs.ts
@@ -1,0 +1,40 @@
+import { applyDecorators } from '@nestjs/common';
+import {
+  ApiBadRequestResponse,
+  ApiOkResponse,
+  ApiOperation,
+  ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
+
+export function ApiLoginAdmin() {
+  return applyDecorators(
+    ApiOperation({
+      summary: `관리자 로그인 API`,
+    }),
+    ApiOkResponse({
+      description: 'Ok',
+      schema: {
+        properties: {
+          message: {
+            type: 'string',
+          },
+        },
+      },
+      example: {
+        message: '로그인이 성공적으로 처리되었습니다.',
+      },
+    }),
+    ApiBadRequestResponse({
+      description: 'Bad Request',
+      example: {
+        message: '오류 메세지',
+      },
+    }),
+    ApiUnauthorizedResponse({
+      description: 'Unauthorized',
+      example: {
+        message: '아이디 혹은 비밀번호가 잘못되었습니다.',
+      },
+    }),
+  );
+}
